ai_assistant/requirements.txt
# Core AI and API dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Google APIs (Gmail only)
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Email handling
email-validator>=2.0.0

# Utilities
requests>=2.31.0
pytz>=2023.3

# Web automation for CMS integration
playwright>=1.40.0

